Use ColorDialog for DialogChangeLayerColor Fix ColorDialog updating input field before the text becomes valid
